+++ CHANGELOG.md
@@ -8,6 +8,8 @@ SPDX-License-Identifier: CC-BY-SA-4.0
 
 ## Unreleased
 
+- Can create groups from the add article and the edit article details forms.
+
 ## 26.02.5
 
 - Add sqlite support and switch to sqlite for production and development.